requirements.txt
# Azure AI Foundry (NEW SDK - preview)
azure-ai-projects>=1.0.0b7
azure-ai-agents>=1.0.0
azure-identity>=1.19.0
azure-ai-evaluation>=1.0.0b6
azure-ai-inference>=1.0.0b9

# MCP Client  
mcp>=1.0.0

# Knowledge Graph
networkx>=3.2

# UI Dashboard
streamlit>=1.40.0
plotly>=5.24.0
streamlit-agraph>=0.0.45

# Utilities
python-dotenv>=1.0.0
pydantic>=2.9.0
httpx>=0.27.0
rich>=13.9.0

# Email
aiosmtplib>=3.0.0

# Testing
pytest>=8.3.0
